renamed menu.c and menu.h to main.c and main.h
